fixed tests

+++ src/rematch/evaluation/algorithm/finditer_algorithm.cpp
@@ -1,4 +1,5 @@
 #include "finditer_algorithm.hpp"
+#include <cstdint>
 
 #ifdef TRACY_ENABLE
 #include <tracy/Tracy.hpp>
@@ -46,7 +47,7 @@ void FinditerAlgorithm::update_output_nodes(ExtendedDetVAState* next_state,
   ZoneScoped;
   #endif
 
-  if (next_state->phase < pos_i_) {
+  if (next_state->phase < (int64_t)pos_i_) {
     next_state->set_phase(pos_i_);
 
     next_state->set_node(next_node);
