+++ src/rematch/output_enumeration/ecs.hpp
@@ -51,7 +51,7 @@ class ECS { // Enumerable Compact Set
    * referring to.
    */
   ECSNode* create_extend_node(ECSNode* node, std::bitset<64> variable_markers,
-                              int64_t document_position);
+                              int document_position);
   /**
    * Union nodes allow us to enumerate multiple outputs starting from a
    * single node.
